requirements.txt
numpy>=1.21.0
opencv-python>=4.5.5.64

# Azure Cognitive Services
azure-cognitiveservices-vision-customvision>=3.1.0
azure-identity>=1.10.0
azure-storage-blob>=12.13.0

# Sinch
sinch>=1.0.0

# Async and utility packages
asyncio>=3.4.3
msrest>=0.7.1